refactor(structure): Flatten one-file schema folders to src/schemas/
Phase 2 of the architectural cleanup. Previously every schema type lived
in its own folder (`src/array/`, `src/bigint/`, etc.) even when the folder
contained only a single `generate<Type>.ts` file. That nesting added no
organizational value — just one extra layer to navigate.

Move:
  src/<type>/generate<Type>.ts → src/schemas/<type>.ts

For the 13 single-file schemas: array, bigint, blob, date, file,
intersect, map, object, record, set, tuple, union, variant.

Multi-file schemas keep their internal layout, just under `src/schemas/`:
  src/string/  → src/schemas/string/   (6 files)
  src/number/  → src/schemas/number/   (3 files)

Net effect: 13 fewer folders. Reduces friction when grepping or scanning
the project layout. All file moves go through `git mv` so blame and
history are preserved.

Bumps Wallaby's discovered coverage to 92.18%.
