+++ src/schemas/variant.ts
